Use the right platform independent ptr type.
git-svn-id: http://root.cern.ch/svn/root/trunk@42175 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/cling/lib/Interpreter/ValuePrinter.cpp
@@ -46,7 +46,8 @@ static void StreamRef(llvm::raw_ostream& o, const void* v) {
 }
   
 static void StreamPtr(llvm::raw_ostream& o, const void* v) {
-  o << *(uint64_t*)v << "\n";
+  o << *(intptr_t
+*)v << "\n";
 }
   
 static void StreamObj(llvm::raw_ostream& o, const void* v) {
